requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.54
psycopg[binary]
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.7.4
python-slugify==8.0.1
python-dotenv==1.1.1